internal heading type filter recognizes tree structure of headings
* a heading is not removed if it has any matched child even if the heading
  itself isn't matched (tree-aware)

* add tree.filter(), which is the filter function that do above
* add tree.remove_child()
* add some accessor functions for the attributes of tree node

+++ autoload/unite/sources/outline/modules/tree.vim
@@ -1,7 +1,7 @@
 "=============================================================================
 " File    : autoload/unite/source/outline/modules/tree.vim
 " Author  : h1mesuke <himesuke@gmail.com>
-" Updated : 2011-03-26
+" Updated : 2011-03-27
 " Version : 0.3.2
 " License : MIT license {{{
 "
@@ -38,26 +38,45 @@ function! s:Tree_append_child(parent, child)
   let a:child.parent = a:parent
 endfunction
 
+function! s:Tree_remove_child(parent, child)
+  call remove(a:parent.children, index(a:parent.children, a:child))
+endfunction
+
+function! s:get_id(node)
+  return get(a:node, 'id', get(a:node, 'source__heading_id'))
+endfunction
+
+function! s:get_parent(node)
+  return get(a:node, 'parent', get(a:node, 'source__heading_parent'))
+endfunction
+
+function! s:get_children(node)
+  return get(a:node, 'children', get(a:node, 'source__heading_children', []))
+endfunction
+
+function! s:is_toplevel(node)
+  return !(has_key(a:node, 'parent') || has_key(a:node, 'source__heading_parent'))
+endfunction
+
+function! s:is_leaf(node)
+  return !(has_key(a:node, 'children') || has_key(a:node, 'source__heading_children'))
+endfunction
+
 function! s:Tree_build(headings)
-  let root = { 'children': [] }
+  let root = { 'level': 0, 'children': [] }
   if empty(a:headings) | return root | endif
 
-  let context   = [{ 'level': -1 }] | " stack
-  let prev_node =  a:headings[0]
-  for node in a:headings
-    while context[-1].level >= node.level
+  let context = [root] | " stack
+  let prev_heading =  a:headings[0]
+  for heading in a:headings
+    while context[-1].level >= heading.level
       call remove(context, -1)
     endwhile
-    if context[-1].level > 0
-      call s:Tree_append_child(context[-1], node)
-    endif
-    call add(context, node)
+    call s:Tree_append_child(context[-1], heading)
+    call add(context, heading)
   endfor
 
-  let is_toplevel = '!has_key(v:val, "parent")'
-  let root.children = filter(copy(a:headings), is_toplevel)
-
-  return root
+  return s:Tree_normalize(root)
 endfunction
 
 function! s:Tree_convert_id_to_ref(candidates)
@@ -70,37 +89,83 @@ function! s:Tree_convert_id_to_ref(candidates)
       let cand.source__heading_parent = cand_table[cand.source__heading_parent]
     endif
     if has_key(cand, 'source__heading_children')
-      let cand.source__heading_children = map(cand.source__heading_children, 'cand_table[v:val]')
+      let cand.source__heading_children =
+            \ map(cand.source__heading_children, 'cand_table[v:val]')
     endif
   endfor
+  return a:candidates
 endfunction
 
 function! s:Tree_convert_ref_to_id(candidates)
+  for cand in a:candidates
+    if has_key(cand, 'source__heading_parent')
+      let cand.source__heading_parent = cand.source__heading_parent.source__heading_id
+    endif
+    if has_key(cand, 'source__heading_children')
+      let cand.source__heading_children =
+            \ map(cand.source__heading_children, 'v:val.source__heading_id')
+    endif
+  endfor
+  return a:candidates
+endfunction
+
+function! s:Tree_filter(treed_list, pred, ...)
+  if empty(a:treed_list) | return a:treed_list | endif
+
+  let do_remove_child = (a:0 ? a:1 : 0)
+
+  let marked = {}
+  for node in filter(copy(a:treed_list), 's:is_toplevel(v:val)')
+    call s:mark(node, a:pred, marked, do_remove_child)
+  endfor
+
+  let treed_list = []
+  for node in a:treed_list
+    if marked[s:get_id(node)]
+      call add(treed_list, node)
+    endif
+  endfor
+  return treed_list
+endfunction
+
+function! s:mark(node, pred, marked, do_remove_child)
+  let child_marked = 0
+  if !s:is_leaf(a:node)
+    for child in s:get_children(a:node)
+      if s:mark(child, a:pred, a:marked, a:do_remove_child)
+        let child_marked = 1
+      elseif a:do_remove_child
+        call s:Tree_remove_child(a:node, child)
+      endif
+    endfor
+  endif
+  let self_marked = (child_marked || a:pred.call(a:node))
+  let a:marked[s:get_id(a:node)] = self_marked
+  return self_marked
 endfunction
 
 function! s:Tree_flatten(tree)
   let headings = []
-  for node in get(a:tree, 'children', [])
-    let node.level = s:is_toplevel(node) ? 1 : node.parent.level + 1
-    call add(headings, node)
-    let headings += s:Tree_flatten(node)
-  endfor
+  if !s:is_leaf(a:tree)
+    for node in s:get_children(a:tree)
+      let node.level = s:is_toplevel(node) ? 1 : node.parent.level + 1
+      call add(headings, node)
+      let headings += s:Tree_flatten(node)
+    endfor
+  endif
   return headings
 endfunction
 
-function! s:is_leaf(node)
-  return !has_key(a:node, 'children')
-endfunction
-
-function! s:is_toplevel(node)
-  return !has_key(a:node, 'parent')
-endfunction
-
 function! s:Tree_normalize(root)
-  " unlink the references to the root node
-  for node in get(a:root, 'children', [])
-    if has_key(node, 'parent') | unlet node.parent | endif
-  endfor
+  call extend(a:root, { 'id': 0, 'level': 0 })
+  if has_key(a:root, 'children')
+    " unlink the references to the root node
+    for node in a:root.children
+      if has_key(node, 'parent')
+        unlet node.parent
+      endif
+    endfor
+  endif
   return a:root
 endfunction
